Seperated Staff Inbox from Staff Messages
Staff Inbox is now seperately named and linked to from the Staff Messages (the user inbox part of the staff pm system). This prevents a bug where staff could send messages to the staff but could then never clear the alert bar. It also makes it plain whether you are in the user inbox or staff inbox which was confused before. ALso fixed a minor bug in user page alt value for avatar.