Added more info to staff pm interface. Fixed overlapping links in header.
Changed Staff PM's to show all unanswered by default (for user level). Added num of each to links and pm status in tables of staff messages. Fixed an issue with the checkboxes for the user_inbox in staff Pm's so they work and are only displayed when appropriate.
Fixed overlapping links under header bar and similar issue with buttons in bbcode helper.
Added button for background [bg] tag which prompts for color. Added a few more named colors to attribs.